[fix] API 응답에서 퍼센티지 데이터 수정 (#261)
* #260 fix: 퍼센티지 데이터를 정수타입에서 Double 타입으로 변경

* #260 test: 테스트 코드 수정

- 실수값 검증하도록 수정

* #260 fix: 메소드의 타입 변경

+++ src/main/java/codesquad/fineants/spring/api/dashboard/service/DashboardService.java
@@ -56,8 +56,8 @@ public class DashboardService {
 		}
 		Integer totalAnnualDividendYield = totalCurrentValuation != 0 ?
 			(int)((totalAnnualDividend / totalCurrentValuation) * 100) : 0;
-		Integer totalGainRate = totalInvestment != 0 ?
-			(int)((totalGain.doubleValue() / totalInvestment.doubleValue()) * 100) : 0;
+		Double totalGainRate = totalInvestment != 0 ?
+			(totalGain.doubleValue() / totalInvestment.doubleValue()) * 100 : 0.0;
 
 		return OverviewResponse.of(member.getNickname(), totalValuation, totalInvestment,
 			totalGain, totalGainRate, totalAnnualDividend, totalAnnualDividendYield);
